Split up filters in smaller components
